requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
motor>=3.3.2
beanie>=1.25.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
email-validator>=2.1.0
